\name{nhl}
\docType{data}
\alias{nhl}

\title{Response Times for 31 Advanced non-Hodgkin's Lymphoma Patients}

\description{
  This data set, which appears in chapter 7 of \cite{Matthews and Farewell}, records the observed time from treatment to either death or end of follow-up 
  for 31 patients suffering from advanced non-Hodgkin's lymphoma.

  \tabular{ll}{
    \bold{Variable name} \tab \bold{description}  \cr
	 time           \tab Observed response measurement, in months  \cr
	 status         \tab  Status, 1 = died, 0 = alive
	 }
	 
}

\usage{nhl}
\format{A data frame consisting of 31 cases, with 2 variables for each case.}

\references{Matthews, D.E. and Farewell, V.T.  (2007) \emph{Using and Understanding Medical Statistics.
4th Completely Enlarged and Revised Edition}.  Basel:  S Karger AG.}

\keyword{datasets}
